requirements.txt
WTForms==3.0.1
Flask==2.2.3
Flask-APScheduler==1.12.4
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.9
openai==0.27.6
boto3==1.26.151
beautifulsoup4==4.12.2
lxml==4.9.3